Update oracle integration test to use OracleReader class
